fix(kernel): stop truncating the assertion lifetime before comparing it
Refuted by Codex (P35-05aR3-02, e8508b0). int(exp) - int(iat) truncated both
sides before subtracting, so a lifetime of up to 300.99 seconds computed as 300
and passed a 300-second ceiling.

It only reproduces when iat is a whole number - which is what every real identity
provider emits. The maker's own probe used a fractional iat taken straight from
datetime.now(), and that masked it completely: 300.9 with a fractional iat is
correctly refused, and with an integer iat it is accepted. A verifier that
constructed the assertion the way a real IdP would found it immediately.

RFC 7519 NumericDate is "a JSON numeric value", not an integer, so fractional
seconds are conformant input rather than an edge case someone had to invent.
Truncating them on the way into a comparison discards precision the specification
guarantees.

Worth recording which way round this one went. The previous seven refutations in
this repository were propositions claiming more than the code did. This time the
proposition was right - "exceeds 300s is refused" - and the code failed to
implement it. That is the first refutation that found a genuine implementation
defect rather than a wording defect, and it is the one that took an
independently-constructed probe to surface.

465 tests. Mutation probe: restoring int() truncation breaks 3.

+++ services/platform-kernel/python/platform_kernel/subject_assertion.py
@@ -252,7 +252,17 @@ def verify_subject_assertion(token: str) -> SubjectClaims:
     # Lifetime ceiling. Checked after the signature holds, so an unsigned assertion is never
     # inspected for its claims.
     try:
-        lifetime = int(claims["exp"]) - int(claims["iat"])
+        # `float`, not `int`. Refuted by Codex 2026-08-01 (`P35-05aR3-02`): truncating both sides
+        # before subtracting let a lifetime of up to 300.99s pass a 300s ceiling, because
+        # `int(iat + 300.9) - int(iat)` is 300 whenever `iat` is a whole number — which is what
+        # every real identity provider emits.
+        #
+        # RFC 7519 NumericDate is "a JSON numeric value", not an integer, so fractional seconds
+        # are conformant input and must not be discarded on the way into a comparison.
+        #
+        # The proposition was right and the code was wrong. Worth recording, because the previous
+        # seven refutations in this repository were the other way round.
+        lifetime = float(claims["exp"]) - float(claims["iat"])
     except (TypeError, ValueError):
         raise AssertionVerificationError(
             "assertion is not valid", reason="non_numeric_lifetime"
